+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `delivery-options` to be rendered as radio
+- Bunch of new tests for radio implementation
+
+### Changed
+
+- Logic to display a radio button instead of checkbox
+
+### Fixed
+
+- CHeckbox behavior acting as radio in some scenarios
+
 ## [3.140.1] - 2025-09-29
 
 ### Added